+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # developing
+- [refactor: ORM Builder pattern that supports select statement](https://github.com/beego/beego/pull/5460)
 - [fix: catch missed error on migration](https://github.com/beego/beego/pull/5455)
 
 # v2.1.2